chore: Update dependencies and add opendal library
This commit updates the dependencies in the project and adds the opendal library. The Cargo.toml files in the root directory and collab-core directory have been modified to include the necessary dependencies. Additionally, the brainrot.yaml file in the rpc/brainrot/etc directory has been updated to whitelist new API endpoints. The SignUpLogic file in the rpc/brainrot/internal/logic/user directory has been modified to set the initial reputation of new users to 10. Finally, the CommentModel file in the model directory has been modified to handle errors when querying for comments by article ID.